chore: move statistical analysis prediction per independent variable to a separate file
